+++ example/demo_india.dart
@@ -1,7 +1,0 @@
-import 'package:tithi_engine/tithi_engine.dart';
-import 'package:tithi_engine/data/india.dart';
-
-void main() {
-  final panchang = Panchang([registerIndia]);
-  print(panchang.tithiOnDate(DateTime(2026, 2, 15), 'Ujjain').displayName);
-}